Add xcode project
finally downloaded it. Add a new empty project, and add the two packages as static libraries.
This seems to be the best way to represent them in a way xcode likes (and kinda makes sense,
more or less how go is treating it). This gives us better analysis tools for all the memory
leaks and bugs, and better autocompletion. We just ignore the actual library output and let
go compile it in as needed